+++ CHANGES.rst
@@ -9,6 +9,7 @@ Improvements:
  * MXKeyBackup: Add MXKeyBackupStateNotTrusted state.
  * MXKeyBackup: Do not reset MXKeyBackup.keyBackupVersion in error states.
  * MXKeyBackup: Implement the true deleteKeyBackupVersion Client-Server API.
+ * Crypto: Cancel share request on restore/import (vector-im/riot-ios/issues/#2232).
 
 Bug Fix:
  * Crypto: Device deduplication method sometimes crashes (vector-im/riot-ios/issues/#2167).